+++ framework/core/__test__/configInjector.test.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
 import { number, SchemaValidator, string } from '@forklaunch/validator/zod';
 import { ConfigInjector } from '../src/services/configInjector';
 import { Lifetime } from '../src/services/types/configInjector.types';