+++ framework/hyper-express/__test__/zod.forklaunch.hyperExpress.test.ts
@@ -246,8 +246,10 @@ describe('handlers', () => {
           name: string
         }
       },
-      async (req, res) => {
-        req.query.name;
+      async (req) => {
+        if (req.query.name) {
+          throw new Error('Name is required');
+        }
       }
     );
     application.use(checkoutMiddleware);